raft: randomized_nemesis_test: network
`network` is a simple priority queue of "events", where an event is a
message associated with delivery time. Each message contains a source,
a destination, and payload. The queue uses a logical clock to decide
when to deliver messages; it delivers are messages whose associated
times are smaller than the current time.

The exact delivery method is unknown to `network` but passed as a
`deliver_t` function in the constructor. The type of payload is generic.

The `deliver_t` function may return `false`, indicating that it failed
to deliver a message so it should be retried later (if it doesn't want a
retry it should return `true`). This may happen when e.g. `deliver_t`
tries to push the message onto a queue which is currently full.