storage_proxy: do not crash on LOCAL_QUORUM access to a DC with zero replication
If a table that is not replicated to a certain DC (rf=0) is accessed
with LOCAL_QUORUM on that DC the current code will crash since the
'targets' array will be empty and read executor does not handle it.
Fix it by replying with empty result.

Fixes #8354

Message-Id: <YGro+l2En3fF80CO@scylladb.com>
(cherry picked from commit cd24dfc7e5b5e83d783d177835c4dbb3b8efef4f)

[avi: re-added virtual keyword when backporting, since
      4.4 and below don't have 020da49c89b5704dd0c2a0f0e58a9906e00ae0b3]