tracing: make client_state.user to be lw_shared_ptr
to save a memory allocation and string formatting during tracing for the
username.

To optimize code hot paths for tracing we are trying to remove dynamic
allocations and memory copying as much as possible. One of the biggest
contributor to the performance overhead in tracing is string formatting.

User authentication object exists in the lifetime context of the
connection. A user does an authentication only once. Thus we can omit
copying this object in the hot path of query execution by reusing
a shared ptr to it.

This patch changes an option to lw_shared_ptr for auth user context to
keep only the ptr instead of the value inside of the trace state params.

Signed-off-by: Ivan Prisyazhnyy <ivan@scylladb.com>