table: introduce trigger_offstrategy_compaction()
this function will be used on repair-based operation completion,
to notify table about the need to start offstrategy compaction
process on the maintenance sstables produced by the operation.
Function which notifies about bootstrap and replace completion
is changed to use this new function.
Removenode and decommission will reuse this function.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>