logalloc: background reclaim: improve tradeoff between efficiency and latency
Reclaim receives the number of bytes it needs to reclaim as a parameter,
and uses it to tune efficiency. If more memory is needed to be reclaimed,
reclaim will evict enough items so that we have sufficient segments to
compact and release the required memory. This up-front eviction increases
the probablilty that some segments will have larger amounts of free space
than the minimum threshold, and so less compaction will be needed.

Synchronous reclaim uses the amount of memory needed for this parameter.
For synchronous reclaim, improving efficiency also improves latency.

Background reclaim uses the difference between the current amount of
free memory amd the free memory goal for this parameter. This is
problematic, since latency-to-first-reclaimed-segment is proportional
to the parameter, and so are the shares given to background reclaim.
As a result, time-to-first-segment can be always in the future. Shares
grow as a result of the backlog growing, but so does the eviction work
we have to do before the first segment is released, and the two effects
cancel out.

Fix by providing a fixed batch size for reclaim. At low backlogs, this
allows reclaim to work more efficiently since the batch size is higher.
We don't care if the amount of free memory is a little lower as a result,
there is no difference between 60 MB free memory and 58 MB free memory.
At higher backlogs, efficiency won't be as high as the existing code,
but at least we'll have bounded time on freeing memory and shares won't
increase indefinitely.