raft: fsm: fix assertion failure on stray rejects
When `raft::fsm` received an `append_reply::rejected` message as a
leader from a follower of the same term, it would assert that
the index of the next log entry to replicate to that follower
(`progress.next_idx`) is not equal to the last known match index of the
follower (`progress_next_idx`).However, this assumption would not
always be true, as the regression test included in this commit shows.

We fix the problem by simply removing the assert. The assert failure comes
from a rejection message which was sent by a follower before an accept
message, but delievered to the leader after that accept message (so the
two messages were reordered in transit); thus the rejection message was
a ``stray'' message that nevertheless passed the `is_stray_reject`
check.

The commit improves the documentation by explaining that
`is_stray_reject` may return false negatives, and `is_stray_reject`
should only be used for optimization (as it currently will be after
we remove the assert), not for safety. It is always safe to process
a reject message and rollback `progress.next_idx` a bit; it may cause
some redundant resending of `append_request`, but that's fine.