test: perf: add a performance test for logalloc
Proving changes under a real workload can be challenging because there
are too many moving parts. So we should have a dedicated framework,
based on code from test/boost/logalloc_test, to recduce cycle times.

This commit contains a model for testing the allocator and a sample test.

The model attempts to be a simplification of logalloc's use case in the
database. It allows "writing" and "reading" values of chosen key and size,
provides an eviction function, and allows shuffling the order of eviction (LRU).

The current implementation of this model uses a vector to store the index of
all keys, which limits possible sizes of the test dataset, but takes tree
operations out the picture. This implementation contains two LSA regions:
evictable "cache" and nonevictable "memtable".

"Writes" are allocated into "memtable". This possibly frees an older entry for
the same key. Freeing from "cache" is called "invalidation" in the stats, and
freeing from "memtable" - "overwrite". If there was no older entry present in
memory, it's a "fresh" write. When "memtable" grows over a threshold, its merged
into "cache".

The difference between "reads" and "writes" is that "reads" to non-present keys
are allocated directly to "cache".

The reason for this two-region model is that regions are made non-compactible
during allocation, so that references are not invalidated when mutliple objects
are allocated in the same region during one logical operation. This means
that allocating to region A can't reclaim memory from region A, only B. Due to
this, a model with one region can't do synchronous reclaim, only background
reclaim.

An alternative model could use only one region and prevent OOMs using
allocating_section.